# Admitix- Root Folder/Test Cases/admitix_api_test_kit/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Admitix API",
    "description": "Production Ready Management API",
    "version": "1.0.0"
  },
  "paths": {
    "/admission-cycles/": {
      "get": {
        "tags": [
          "Admission Cycles"
        ],
        "summary": "Get Admission Cycles Route",
        "operationId": "get_admission_cycles_route_admission_cycles__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdmissionCycleRead"
                  },
                  "type": "array",
                  "title": "Response Get Admission Cycles Route Admission Cycles  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admission Cycles"
        ],
        "summary": "Create Admission Cycle Route",
        "operationId": "create_admission_cycle_route_admission_cycles__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdmissionCycleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionCycleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admission-cycles/{admission_cycle_id}": {
      "get": {
        "tags": [
          "Admission Cycles"
        ],
        "summary": "Get Admission Cycle By Id Route",
        "operationId": "get_admission_cycle_by_id_route_admission_cycles__admission_cycle_id__get",
        "parameters": [
          {
            "name": "admission_cycle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Admission Cycle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionCycleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Admission Cycles"
        ],
        "summary": "Update Admission Cycle Route",
        "operationId": "update_admission_cycle_route_admission_cycles__admission_cycle_id__patch",
        "parameters": [
          {
            "name": "admission_cycle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Admission Cycle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdmissionCycleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionCycleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admission Cycles"
        ],
        "summary": "Delete Admission Cycle Route",
        "operationId": "delete_admission_cycle_route_admission_cycles__admission_cycle_id__delete",
        "parameters": [
          {
            "name": "admission_cycle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Admission Cycle Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ai-verifications/": {
      "get": {
        "tags": [
          "AI Verification"
        ],
        "summary": "Get Ai Verifications Route",
        "operationId": "get_ai_verifications_route_ai_verifications__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AIVerificationRead"
                  },
                  "type": "array",
                  "title": "Response Get Ai Verifications Route Ai Verifications  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AI Verification"
        ],
        "summary": "Create Ai Verification Route",
        "operationId": "create_ai_verification_route_ai_verifications__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIVerificationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIVerificationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ai-verifications/{verification_id}": {
      "get": {
        "tags": [
          "AI Verification"
        ],
        "summary": "Get Ai Verification By Id Route",
        "operationId": "get_ai_verification_by_id_route_ai_verifications__verification_id__get",
        "parameters": [
          {
            "name": "verification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Verification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIVerificationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "AI Verification"
        ],
        "summary": "Update Ai Verification Route",
        "operationId": "update_ai_verification_route_ai_verifications__verification_id__patch",
        "parameters": [
          {
            "name": "verification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Verification Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIVerificationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIVerificationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AI Verification"
        ],
        "summary": "Delete Ai Verification Route",
        "operationId": "delete_ai_verification_route_ai_verifications__verification_id__delete",
        "parameters": [
          {
            "name": "verification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Verification Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIVerificationCreate": {
        "properties": {
          "ocr_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Text"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "blur_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blur Score"
          },
          "missing_fields": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missing Fields"
          },
          "name_match": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Match"
          },
          "status": {
            "$ref": "#/components/schemas/AIVerificationStatus",
            "default": "pending"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "AIVerificationCreate",
        "description": "Payload used by the AI/OCR pipeline to record a verification\nresult against a document."
      },
      "AIVerificationRead": {
        "properties": {
          "ocr_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Text"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "blur_score": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blur Score"
          },
          "missing_fields": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missing Fields"
          },
          "name_match": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Match"
          },
          "status": {
            "$ref": "#/components/schemas/AIVerificationStatus",
            "default": "pending"
          },
          "verification_id": {
            "type": "string",
            "format": "uuid",
            "title": "Verification Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "type": "object",
        "required": [
          "verification_id",
          "document_id",
          "verified_at"
        ],
        "title": "AIVerificationRead"
      },
      "AIVerificationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "passed",
          "failed",
          "manual_review"
        ],
        "title": "AIVerificationStatus"
      },
      "AIVerificationUpdate": {
        "properties": {
          "ocr_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Text"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "blur_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blur Score"
          },
          "missing_fields": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missing Fields"
          },
          "name_match": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Match"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIVerificationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "type": "object",
        "title": "AIVerificationUpdate"
      },
      "AdmissionCycleCreate": {
        "properties": {
          "academic_year": {
            "type": "string",
            "maxLength": 20,
            "title": "Academic Year",
            "examples": [
              "2026-27"
            ]
          },
          "application_start": {
            "type": "string",
            "format": "date",
            "title": "Application Start"
          },
          "application_end": {
            "type": "string",
            "format": "date",
            "title": "Application End"
          },
          "status": {
            "$ref": "#/components/schemas/AdmissionCycleStatus",
            "default": "upcoming"
          },
          "institution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Institution Id"
          }
        },
        "type": "object",
        "required": [
          "academic_year",
          "application_start",
          "application_end",
          "institution_id"
        ],
        "title": "AdmissionCycleCreate"
      },
      "AdmissionCycleRead": {
        "properties": {
          "academic_year": {
            "type": "string",
            "maxLength": 20,
            "title": "Academic Year",
            "examples": [
              "2026-27"
            ]
          },
          "application_start": {
            "type": "string",
            "format": "date",
            "title": "Application Start"
          },
          "application_end": {
            "type": "string",
            "format": "date",
            "title": "Application End"
          },
          "status": {
            "$ref": "#/components/schemas/AdmissionCycleStatus",
            "default": "upcoming"
          },
          "cycle_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cycle Id"
          },
          "institution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Institution Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "academic_year",
          "application_start",
          "application_end",
          "cycle_id",
          "institution_id",
          "created_at"
        ],
        "title": "AdmissionCycleRead"
      },
      "AdmissionCycleStatus": {
        "type": "string",
        "enum": [
          "upcoming",
          "open",
          "closed",
          "archived"
        ],
        "title": "AdmissionCycleStatus"
      },
      "AdmissionCycleUpdate": {
        "properties": {
          "academic_year": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Academic Year"
          },
          "application_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Start"
          },
          "application_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application End"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdmissionCycleStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "AdmissionCycleUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}